Move get_terminal_size to pip._internal.compat
